Safer timezone handling
`datetime` is much modern than `time`, but to handle local timezone we must
either use `time` or add another external dependency (such as pytz).

OldestDate was set to year 1900, which caused time.mktime() to fail on some
platforms which can't handle dates earlier than 1970 (like i686). OldestDate is
now set to the date with timestamp 0, which should be much safer.

Additionally, OverflowError's are now caught, logged, and the corresponding
timestamp is set to 0.

Fixes #9.