Ignore invalid channel names
Previously, trying to display a channel with an invalid JID (either missing user
part, such as @example.com, or missing domain, such as user@) resulted in Urwid
trying to render text widget with zero width, which failed and result in a crash
because of an AssertionError.

But trying to handle these invalid JIDs makes no sense anyway.

These JIDs are now detected and rejected by the client, so they don't cause the
UI to crash anymore.

A future patch will add a visual notification for such invalid channel names.

Many *many* thanks to Simon (@imaginator) for his help in finding this (and
sorry for the delay).