Scroll up when a new message is posted to a channel
When new content is available, try to display it while keeping the focused item
on screen. If not possible, the user is notified with a message in the status
bar.

This also introduces a new function, UI.safe_status_set_text(), to change the
status bar safely from another thread, a callback, or an Urwid internal
function.

Fixes #7.